Use existing DependencyVersions object
